requirements.txt
streamlit
transformers
torch
sentencepiece
pdfplumber
python-docx
sumy
nltk
pytesseract
Pillow
